requirements.txt
# Voice Recognition
SpeechRecognition==3.10.0
pyaudio==0.2.14
pyttsx3==2.90
pyobjc>=9.0; sys_platform == "darwin"

# Calendar Integration
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0

# Natural Language Processing
groq>=0.34.1
dateparser==1.2.0

# Location and Timezone
pytz>=2023.3
geopy>=2.4.0
timezonefinder>=6.2.0
geographiclib>=2.0.0

# Utilities
python-dotenv==1.0.0
pydantic>=2.5.0

# Web Framework
flask>=3.0.0
flask-cors>=4.0.0
gunicorn>=21.2.0

# Text-to-Speech
elevenlabs>=1.0.0

